refactor: split services and delete console logs
