+++ backend/Program.cs
@@ -11,6 +11,7 @@ builder.Services.AddScoped<ContactService>();
 builder.Services.AddScoped<HotelService>();
 builder.Services.AddScoped<RoomService>();
 builder.Services.AddScoped<RoomFiltersService>();
+builder.Services.AddScoped<RoomInfoServices>();
 builder.Services.AddScoped<RoomTemplateService>();
 builder.Services.AddScoped<ServiceService>();
 builder.Services.AddScoped<UserService>();
